Change `Alice` to `tumbler`
In the current version, I always get confused who is the tumbler, either Alice or Bob. The two are not peers [Bob knows more than Alice] and the two have different tasks, so I think it is better to differentiate them more clearly. In this PR I call `Alice` the `tumbler service provider` and `Bob` and `Carol` are the `users` and I replace most of her name with `the tumbler` so that it is always clear who has which role. In my opinion, this makes it a lot easier to conceptualize.